+++ src/domain/cart/components/cart-layout/LabelPriceContainer.tsx
@@ -1,19 +1,15 @@
 import styled from "@emotion/styled";
-import { getShoppingCartData } from "../../../../api/cart";
 import Flex from "../../../../components/common/styled/Flex";
 import InfoText from "../../../../components/common/text/InfoText";
 import LabelPrice from "../../../../components/common/text/LabelPrice";
-import { useAPIDataContext } from "../../../../context/APIDataProvider";
-import { useOrderListContext } from "../../../order/context/OrderListProvider";
 import { formatKRWString } from "../../../../utils/formatKRWString";
-import { FREE_SHIPPING_STANDARD } from "../../../order/hooks/OrderConstants";
+import { useOrderListContext } from "../../../order/context/OrderListProvider";
+import { FREE_SHIPPING_STANDARD } from "../../../order/consts/OrderConstants";
+import { useCartAPIData } from "../../hooks/useCartAPIData";
 import { calculateOrders } from "../../../order/utils/calculateOrders";
 
 const LabelPriceContainer = () => {
-  const { data: cartListData } = useAPIDataContext({
-    fetcher: getShoppingCartData,
-    name: "cart",
-  });
+  const { cartListData } = useCartAPIData();
   const { selectedCartItems } = useOrderListContext(cartListData);
 
   const { shippingFee, totalPrice, totalCartPrice } =
