+++ src/domain/order/components/pay-contents/DeliveryInfo.tsx
@@ -1,18 +1,14 @@
 import styled from "@emotion/styled";
-import { getShoppingCartData } from "../../../../api/cart";
 import { Flex } from "../../../../components/common";
 import CheckboxLabel from "../../../../components/common/inputs/CheckboxLabel";
-import { useAPIDataContext } from "../../../../context/APIDataProvider";
-import { useOrderListContext } from "../../context/OrderListProvider";
 import { useCouponContext } from "../../../../pages/order-confirm/context/CouponProvider";
+import { useOrderListContext } from "../../context/OrderListProvider";
+import { useCartAPIData } from "../../../cart/hooks/useCartAPIData";
 import { calculateOrders } from "../../utils/calculateOrders";
 import OrderLabelPridce from "../order-contents/OrderLabelPrice";
 
 function DeliveryInfo() {
-  const { data: cartListData } = useAPIDataContext({
-    fetcher: getShoppingCartData,
-    name: "cart",
-  });
+  const { cartListData } = useCartAPIData();
   const { selectedCartItems, isIsland, handleIsIslandToggle } =
     useOrderListContext(cartListData);
   const { selectedCoupons } = useCouponContext();
